[CDAP-14377] Gets cdap config from cConf, instead of API call - Adds ui.theme.file to cdap.json development config, so we don't rely on API call at all
