+++ cdap-ui/server/config/development/cdap.json
@@ -10,5 +10,6 @@
   "dashboard.bind.port": "11011",
   "enable.preview": "true",
   "dashboard.router.check.timeout.secs": "0",
-  "market.base.url": "http://market.cask.co/dev/v2"
+  "market.base.url": "http://market.cask.co/dev/v2",
+  "ui.theme.file": "cdap-ui/server/config/themes/default.json"
 }
